functional class implementation
