+++ src/main/java/com/waytta/clientinterface/LocalClient.java
@@ -2,6 +2,11 @@ package com.waytta.clientinterface;
 
 import hudson.Extension;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.QueryParameter;
+import hudson.util.FormValidation;
+import org.jenkinsci.Symbol;
+
+import com.waytta.Utils;
 
 public class LocalClient extends BasicClient {
     public static final int DEFAULT_JOB_POLL_TIME = 10;
@@ -34,6 +39,22 @@ public class LocalClient extends BasicClient {
         return targetType;
     }
     
-    //@Extension
-    //public static final BasicClientDescriptor DESCRIPTOR = new BasicClientDescriptor(LocalClient.class);
+    @Symbol("local")
+    public static final class DescriptorImpl extends BasicClientDescriptor {
+        private DescriptorImpl(Class<? extends BasicClient> clazz) {
+            super(clazz);
+        }
+        
+        @Override
+        public String getDisplayName() {
+        	return "local";
+        }
+    	
+        public FormValidation doCheckTarget(@QueryParameter String value) {
+            return Utils.validateFormStringField(value, "Please specify a salt target", "Isn't it too short?");
+        }
+    }
+    
+    @Extension
+    public static final BasicClientDescriptor DESCRIPTOR = new DescriptorImpl(LocalClient.class);
 }
