+++ src/main/resources/com/waytta/clientinterface/LocalBatchClient/config.jelly
@@ -21,6 +21,4 @@
       <f:option value="compound" selected="${instance.getTargetType() == 'compound'}">compound</f:option>
     </select>
   </f:entry>
-
-  <!-- <f:class-entry clazz="com.waytta.clientinterface.BasicClient$LocalBatchClient" /> -->
 </j:jelly>
